+++ pkg/velero/backupManager_test.go
@@ -5,9 +5,9 @@ import (
 	"testing"
 	"time"
 
+	"github.com/cloudogu/k8s-backup-operator/pkg/annotations"
 	corev1 "k8s.io/api/core/v1"
 	"k8s.io/apimachinery/pkg/fields"
-	"k8s.io/apimachinery/pkg/types"
 	"k8s.io/apimachinery/pkg/watch"
 	"sigs.k8s.io/controller-runtime/pkg/client"
 
@@ -43,7 +43,7 @@ func Test_backupManager_CreateBackup(t *testing.T) {
 
 		volumeFsBackup := false
 		expectedVeleroBackup := &velerov1.Backup{
-			ObjectMeta: metav1.ObjectMeta{Name: "testBackup", Namespace: testNamespace, Labels: map[string]string{"app": "ces", "k8s.cloudogu.com/part-of": "backup"}},
+			ObjectMeta: metav1.ObjectMeta{Name: "testBackup", Namespace: testNamespace, Labels: map[string]string{"app": "ces", "k8s.cloudogu.com/part-of": "backup"}, Annotations: map[string]string{}},
 			Spec: velerov1.BackupSpec{
 				IncludedResources:        []string{"configmaps", "secrets", "persistentvolumeclaims", "dogus.k8s.cloudogu.com"},
 				OrLabelSelectors:         selectors,
@@ -79,7 +79,7 @@ func Test_backupManager_CreateBackup(t *testing.T) {
 
 		volumeFsBackup := false
 		expectedVeleroBackup := &velerov1.Backup{
-			ObjectMeta: metav1.ObjectMeta{Name: "testBackup", Namespace: testNamespace, Labels: map[string]string{"app": "ces", "k8s.cloudogu.com/part-of": "backup"}},
+			ObjectMeta: metav1.ObjectMeta{Name: "testBackup", Namespace: testNamespace, Labels: map[string]string{"app": "ces", "k8s.cloudogu.com/part-of": "backup"}, Annotations: map[string]string{}},
 			Spec: velerov1.BackupSpec{
 				IncludedResources:        []string{"configmaps", "secrets", "persistentvolumeclaims", "dogus.k8s.cloudogu.com"},
 				OrLabelSelectors:         selectors,
@@ -119,7 +119,7 @@ func Test_backupManager_CreateBackup(t *testing.T) {
 
 		volumeFsBackup := false
 		expectedVeleroBackup := &velerov1.Backup{
-			ObjectMeta: metav1.ObjectMeta{Name: "testBackup", Namespace: testNamespace, Labels: map[string]string{"app": "ces", "k8s.cloudogu.com/part-of": "backup"}},
+			ObjectMeta: metav1.ObjectMeta{Name: "testBackup", Namespace: testNamespace, Labels: map[string]string{"app": "ces", "k8s.cloudogu.com/part-of": "backup"}, Annotations: map[string]string{}},
 			Spec: velerov1.BackupSpec{
 				IncludedResources:        []string{"configmaps", "secrets", "persistentvolumeclaims", "dogus.k8s.cloudogu.com"},
 				OrLabelSelectors:         selectors,
@@ -174,7 +174,7 @@ func Test_backupManager_CreateBackup(t *testing.T) {
 
 		volumeFsBackup := false
 		expectedVeleroBackup := &velerov1.Backup{
-			ObjectMeta: metav1.ObjectMeta{Name: "testBackup", Namespace: testNamespace, Labels: map[string]string{"app": "ces", "k8s.cloudogu.com/part-of": "backup"}},
+			ObjectMeta: metav1.ObjectMeta{Name: "testBackup", Namespace: testNamespace, Labels: map[string]string{"app": "ces", "k8s.cloudogu.com/part-of": "backup"}, Annotations: map[string]string{}},
 			Spec: velerov1.BackupSpec{
 				IncludedResources:        []string{"configmaps", "secrets", "persistentvolumeclaims", "dogus.k8s.cloudogu.com"},
 				OrLabelSelectors:         selectors,
@@ -230,7 +230,7 @@ func Test_backupManager_CreateBackup(t *testing.T) {
 
 		volumeFsBackup := false
 		expectedVeleroBackup := &velerov1.Backup{
-			ObjectMeta: metav1.ObjectMeta{Name: "testBackup", Namespace: testNamespace, Labels: map[string]string{"app": "ces", "k8s.cloudogu.com/part-of": "backup"}},
+			ObjectMeta: metav1.ObjectMeta{Name: "testBackup", Namespace: testNamespace, Labels: map[string]string{"app": "ces", "k8s.cloudogu.com/part-of": "backup"}, Annotations: map[string]string{}},
 			Spec: velerov1.BackupSpec{
 				IncludedResources:        []string{"configmaps", "secrets", "persistentvolumeclaims", "dogus.k8s.cloudogu.com"},
 				OrLabelSelectors:         selectors,
@@ -278,7 +278,14 @@ func Test_backupManager_CreateBackup(t *testing.T) {
 	})
 	t.Run("should succeed when velero backup is completed", func(t *testing.T) {
 		// given
-		testBackup := &backupv1.Backup{ObjectMeta: metav1.ObjectMeta{Name: "testBackup", Namespace: testNamespace}}
+		testBackup := &backupv1.Backup{ObjectMeta: metav1.ObjectMeta{
+			Name:      "testBackup",
+			Namespace: testNamespace,
+			Annotations: map[string]string{
+				annotations.BlueprintIdAnnotation: "test-blueprint-id",
+				annotations.DogusAnnotation:       "test-dogus",
+			},
+		}}
 
 		mockRecorder := newMockEventRecorder(t)
 		mockRecorder.EXPECT().Event(testBackup, "Normal", "Creation", "Using velero as backup provider")
@@ -286,7 +293,15 @@ func Test_backupManager_CreateBackup(t *testing.T) {
 
 		volumeFsBackup := false
 		expectedVeleroBackup := &velerov1.Backup{
-			ObjectMeta: metav1.ObjectMeta{Name: "testBackup", Namespace: testNamespace, Labels: map[string]string{"app": "ces", "k8s.cloudogu.com/part-of": "backup"}},
+			ObjectMeta: metav1.ObjectMeta{
+				Name:      "testBackup",
+				Namespace: testNamespace,
+				Labels:    map[string]string{"app": "ces", "k8s.cloudogu.com/part-of": "backup"},
+				Annotations: map[string]string{
+					annotations.BlueprintIdAnnotation: "test-blueprint-id",
+					annotations.DogusAnnotation:       "test-dogus",
+				},
+			},
 			Spec: velerov1.BackupSpec{
 				IncludedResources:        []string{"configmaps", "secrets", "persistentvolumeclaims", "dogus.k8s.cloudogu.com"},
 				OrLabelSelectors:         selectors,
@@ -349,13 +364,10 @@ func Test_provider_DeleteBackup(t *testing.T) {
 
 		mockK8sWatchClient := newMockK8sWatchClient(t)
 		mockK8sWatchClient.EXPECT().Create(testCtx, expectedRequest).Return(nil)
-		mockK8sWatchClient.EXPECT().Watch(testCtx, &velerov1.BackupList{},
+		mockK8sWatchClient.EXPECT().Watch(testCtx, &velerov1.DeleteBackupRequestList{},
 			&client.ListOptions{FieldSelector: fields.ParseSelectorOrDie("metadata.name=backup"),
 				Namespace: testNamespace, Raw: &metav1.ListOptions{TimeoutSeconds: &deleteWaitTimeout}}).
 			Return(watchMock, nil)
-		mockK8sWatchClient.EXPECT().Get(testCtx, backup.GetNamespacedName(), expectedRequest).Run(func(ctx context.Context, key types.NamespacedName, obj client.Object, opts ...client.GetOption) {
-			*obj.(*velerov1.DeleteBackupRequest) = *expectedRequestProcessed
-		}).Return(nil)
 
 		recorderMock := newMockEventRecorder(t)
 		recorderMock.EXPECT().Event(backup, corev1.EventTypeNormal, backupv1.ProviderDeleteEventReason, "Trigger velero provider to delete backup.")
@@ -363,11 +375,97 @@ func Test_provider_DeleteBackup(t *testing.T) {
 
 		sut := defaultBackupManager{recorder: recorderMock, k8sClient: mockK8sWatchClient}
 
-		watchTimer := time.NewTimer(time.Second * 2)
+		watchTimer := time.NewTimer(time.Second * 1)
 		go func() {
 			<-watchTimer.C
 			event := watch.Event{Type: watch.Modified, Object: expectedRequestProcessed}
 			watchChannel <- event
+		}()
+
+		// when
+		err := sut.DeleteBackup(testCtx, backup)
+
+		// then
+		require.NoError(t, err)
+	})
+
+	t.Run("should succeed if DeleteBackupRequest is deleted", func(t *testing.T) {
+		// given
+		backup := &backupv1.Backup{ObjectMeta: metav1.ObjectMeta{Name: "backup", Namespace: testNamespace}}
+
+		watchChannel := make(chan watch.Event)
+		watchMock := newMockWatchInterface(t)
+		watchMock.EXPECT().ResultChan().Return(watchChannel)
+		watchMock.EXPECT().Stop()
+
+		expectedRequest := getVeleroDeleteBackupRequest(backup.Name, testNamespace)
+		expectedRequestProcessed := getVeleroDeleteBackupRequest(backup.Name, testNamespace)
+		expectedRequestProcessed.Status.Phase = velerov1.DeleteBackupRequestPhaseProcessed
+
+		mockK8sWatchClient := newMockK8sWatchClient(t)
+		mockK8sWatchClient.EXPECT().Create(testCtx, expectedRequest).Return(nil)
+		mockK8sWatchClient.EXPECT().Watch(testCtx, &velerov1.DeleteBackupRequestList{},
+			&client.ListOptions{FieldSelector: fields.ParseSelectorOrDie("metadata.name=backup"),
+				Namespace: testNamespace, Raw: &metav1.ListOptions{TimeoutSeconds: &deleteWaitTimeout}}).
+			Return(watchMock, nil)
+
+		recorderMock := newMockEventRecorder(t)
+		recorderMock.EXPECT().Event(backup, corev1.EventTypeNormal, backupv1.ProviderDeleteEventReason, "Trigger velero provider to delete backup.")
+		recorderMock.EXPECT().Event(backup, corev1.EventTypeNormal, backupv1.ProviderDeleteEventReason, "Provider delete request successful.")
+
+		sut := defaultBackupManager{recorder: recorderMock, k8sClient: mockK8sWatchClient}
+
+		watchTimer := time.NewTimer(time.Second * 1)
+		go func() {
+			<-watchTimer.C
+			event := watch.Event{Type: watch.Deleted, Object: expectedRequestProcessed}
+			watchChannel <- event
+		}()
+
+		// when
+		err := sut.DeleteBackup(testCtx, backup)
+
+		// then
+		require.NoError(t, err)
+	})
+
+	t.Run("should continue for wrong types and non completed requests", func(t *testing.T) {
+		// given
+		backup := &backupv1.Backup{ObjectMeta: metav1.ObjectMeta{Name: "backup", Namespace: testNamespace}}
+
+		watchChannel := make(chan watch.Event)
+		watchMock := newMockWatchInterface(t)
+		watchMock.EXPECT().ResultChan().Return(watchChannel)
+		watchMock.EXPECT().Stop()
+
+		expectedRequest := getVeleroDeleteBackupRequest(backup.Name, testNamespace)
+		expectedRequestProcessed := getVeleroDeleteBackupRequest(backup.Name, testNamespace)
+		expectedRequestProcessed.Status.Phase = velerov1.DeleteBackupRequestPhaseProcessed
+
+		mockK8sWatchClient := newMockK8sWatchClient(t)
+		mockK8sWatchClient.EXPECT().Create(testCtx, expectedRequest).Return(nil)
+		mockK8sWatchClient.EXPECT().Watch(testCtx, &velerov1.DeleteBackupRequestList{},
+			&client.ListOptions{FieldSelector: fields.ParseSelectorOrDie("metadata.name=backup"),
+				Namespace: testNamespace, Raw: &metav1.ListOptions{TimeoutSeconds: &deleteWaitTimeout}}).
+			Return(watchMock, nil)
+
+		recorderMock := newMockEventRecorder(t)
+		recorderMock.EXPECT().Event(backup, corev1.EventTypeNormal, backupv1.ProviderDeleteEventReason, "Trigger velero provider to delete backup.")
+		recorderMock.EXPECT().Event(backup, corev1.EventTypeNormal, backupv1.ProviderDeleteEventReason, "Provider delete request successful.")
+
+		sut := defaultBackupManager{recorder: recorderMock, k8sClient: mockK8sWatchClient}
+
+		watchTimer := time.NewTimer(time.Second * 1)
+		go func() {
+			<-watchTimer.C
+			event1 := watch.Event{Type: watch.Modified, Object: &velerov1.Backup{}}
+			watchChannel <- event1
+			event2 := watch.Event{Type: watch.Modified, Object: expectedRequest}
+			watchChannel <- event2
+			event3 := watch.Event{Type: watch.Error, Object: expectedRequest}
+			watchChannel <- event3
+			event4 := watch.Event{Type: watch.Modified, Object: expectedRequestProcessed}
+			watchChannel <- event4
 		}()
 
 		// when
@@ -408,7 +506,7 @@ func Test_provider_DeleteBackup(t *testing.T) {
 
 		mockK8sWatchClient := newMockK8sWatchClient(t)
 		mockK8sWatchClient.EXPECT().Create(testCtx, expectedRequest).Return(nil)
-		mockK8sWatchClient.EXPECT().Watch(testCtx, &velerov1.BackupList{},
+		mockK8sWatchClient.EXPECT().Watch(testCtx, &velerov1.DeleteBackupRequestList{},
 			&client.ListOptions{FieldSelector: fields.ParseSelectorOrDie("metadata.name=backup"),
 				Namespace: testNamespace, Raw: &metav1.ListOptions{TimeoutSeconds: &deleteWaitTimeout}}).
 			Return(nil, assert.AnError)
@@ -427,13 +525,14 @@ func Test_provider_DeleteBackup(t *testing.T) {
 		require.ErrorContains(t, err, "failed to create watch for delete backup request backup")
 	})
 
-	t.Run("should return error on get error after watch was successful", func(t *testing.T) {
+	t.Run("should return timeout error after watch channel closes without processed state", func(t *testing.T) {
 		// given
 		backup := &backupv1.Backup{ObjectMeta: metav1.ObjectMeta{Name: "backup", Namespace: testNamespace}}
 
 		watchChannel := make(chan watch.Event)
 		watchMock := newMockWatchInterface(t)
 		watchMock.EXPECT().ResultChan().Return(watchChannel)
+		watchMock.EXPECT().Stop()
 
 		expectedRequest := getVeleroDeleteBackupRequest(backup.Name, testNamespace)
 		expectedRequestProcessed := getVeleroDeleteBackupRequest(backup.Name, testNamespace)
@@ -441,14 +540,15 @@ func Test_provider_DeleteBackup(t *testing.T) {
 
 		mockK8sWatchClient := newMockK8sWatchClient(t)
 		mockK8sWatchClient.EXPECT().Create(testCtx, expectedRequest).Return(nil)
-		mockK8sWatchClient.EXPECT().Watch(testCtx, &velerov1.BackupList{},
+		mockK8sWatchClient.EXPECT().Delete(testCtx, expectedRequest).Return(nil)
+		mockK8sWatchClient.EXPECT().Watch(testCtx, &velerov1.DeleteBackupRequestList{},
 			&client.ListOptions{FieldSelector: fields.ParseSelectorOrDie("metadata.name=backup"),
 				Namespace: testNamespace, Raw: &metav1.ListOptions{TimeoutSeconds: &deleteWaitTimeout}}).
 			Return(watchMock, nil)
-		mockK8sWatchClient.EXPECT().Get(testCtx, backup.GetNamespacedName(), expectedRequest).Return(assert.AnError)
 
 		recorderMock := newMockEventRecorder(t)
 		recorderMock.EXPECT().Event(backup, corev1.EventTypeNormal, backupv1.ProviderDeleteEventReason, "Trigger velero provider to delete backup.")
+		recorderMock.EXPECT().Event(backup, corev1.EventTypeWarning, backupv1.ProviderDeleteEventReason, "Cleanup velero delete request.")
 
 		sut := defaultBackupManager{recorder: recorderMock, k8sClient: mockK8sWatchClient}
 
@@ -463,8 +563,7 @@ func Test_provider_DeleteBackup(t *testing.T) {
 
 		// then
 		require.Error(t, err)
-		require.ErrorIs(t, err, assert.AnError)
-		require.ErrorContains(t, err, "failed to get delete backup request backup")
+		require.ErrorContains(t, err, "failed to delete backup backup: timeout waiting for backup delete request backup")
 	})
 
 	t.Run("should return error and clean up if request status is not processed after timeout", func(t *testing.T) {
@@ -474,16 +573,16 @@ func Test_provider_DeleteBackup(t *testing.T) {
 		watchChannel := make(chan watch.Event)
 		watchMock := newMockWatchInterface(t)
 		watchMock.EXPECT().ResultChan().Return(watchChannel)
+		watchMock.EXPECT().Stop()
 
 		expectedRequest := getVeleroDeleteBackupRequest(backup.Name, testNamespace)
 
 		mockK8sWatchClient := newMockK8sWatchClient(t)
 		mockK8sWatchClient.EXPECT().Create(testCtx, expectedRequest).Return(nil)
-		mockK8sWatchClient.EXPECT().Watch(testCtx, &velerov1.BackupList{},
+		mockK8sWatchClient.EXPECT().Watch(testCtx, &velerov1.DeleteBackupRequestList{},
 			&client.ListOptions{FieldSelector: fields.ParseSelectorOrDie("metadata.name=backup"),
 				Namespace: testNamespace, Raw: &metav1.ListOptions{TimeoutSeconds: &deleteWaitTimeout}}).
 			Return(watchMock, nil)
-		mockK8sWatchClient.EXPECT().Get(testCtx, backup.GetNamespacedName(), expectedRequest).Return(nil)
 		mockK8sWatchClient.EXPECT().Delete(testCtx, expectedRequest).Return(nil)
 
 		recorderMock := newMockEventRecorder(t)
@@ -503,7 +602,7 @@ func Test_provider_DeleteBackup(t *testing.T) {
 
 		// then
 		require.Error(t, err)
-		require.ErrorContains(t, err, "failed to delete backup backup: timout waiting for backup delete request backup")
+		require.ErrorContains(t, err, "failed to delete backup backup: timeout waiting for backup delete request backup")
 	})
 
 	t.Run("should return all errors in multierror and clean up if request status has errors", func(t *testing.T) {
@@ -513,6 +612,7 @@ func Test_provider_DeleteBackup(t *testing.T) {
 		watchChannel := make(chan watch.Event)
 		watchMock := newMockWatchInterface(t)
 		watchMock.EXPECT().ResultChan().Return(watchChannel)
+		watchMock.EXPECT().Stop()
 
 		expectedRequest := getVeleroDeleteBackupRequest(backup.Name, testNamespace)
 		expectedRequestProcessed := getVeleroDeleteBackupRequest(backup.Name, testNamespace)
@@ -521,25 +621,23 @@ func Test_provider_DeleteBackup(t *testing.T) {
 
 		mockK8sWatchClient := newMockK8sWatchClient(t)
 		mockK8sWatchClient.EXPECT().Create(testCtx, expectedRequest).Return(nil)
-		mockK8sWatchClient.EXPECT().Watch(testCtx, &velerov1.BackupList{},
+		mockK8sWatchClient.EXPECT().Watch(testCtx, &velerov1.DeleteBackupRequestList{},
 			&client.ListOptions{FieldSelector: fields.ParseSelectorOrDie("metadata.name=backup"),
 				Namespace: testNamespace, Raw: &metav1.ListOptions{TimeoutSeconds: &deleteWaitTimeout}}).
 			Return(watchMock, nil)
-		mockK8sWatchClient.EXPECT().Get(testCtx, backup.GetNamespacedName(), expectedRequest).Run(func(ctx context.Context, key types.NamespacedName, obj client.Object, opts ...client.GetOption) {
-			*obj.(*velerov1.DeleteBackupRequest) = *expectedRequestProcessed
-		}).Return(nil)
 		mockK8sWatchClient.EXPECT().Delete(testCtx, expectedRequestProcessed).Return(nil)
 
 		recorderMock := newMockEventRecorder(t)
 		recorderMock.EXPECT().Event(backup, corev1.EventTypeNormal, backupv1.ProviderDeleteEventReason, "Trigger velero provider to delete backup.")
 		recorderMock.EXPECT().Event(backup, corev1.EventTypeWarning, backupv1.ProviderDeleteEventReason, "Cleanup velero delete request.")
-		recorderMock.EXPECT().Event(backup, corev1.EventTypeWarning, backupv1.ErrorOnProviderDeleteEventReason, "velero backup delete request error: error1\nvelero backup delete request error: error2")
+		recorderMock.EXPECT().Event(backup, corev1.EventTypeWarning, backupv1.ErrorOnProviderDeleteEventReason, "velero DeleteBackupRequest error: error1\nvelero DeleteBackupRequest error: error2")
 		sut := defaultBackupManager{recorder: recorderMock, k8sClient: mockK8sWatchClient}
 
 		watchTimer := time.NewTimer(time.Second * 1)
 		go func() {
 			<-watchTimer.C
-			close(watchChannel)
+			event := watch.Event{Type: watch.Modified, Object: expectedRequestProcessed}
+			watchChannel <- event
 		}()
 
 		// when